refactor: change component name
