+++ src/widgets/BookigHistory/BookingRow/index.ts
@@ -1,2 +1,2 @@
 export type { TBookingData } from './typing';
-export { Booking } from './ui';
+export { BookingRow } from './ui';
